Refract Model
Creating a separate model for each fragment leads to duplication of fields (user, currentUser and others). It is better to use one MainModel instead of a unique model for each fragment. MainModel now has different types of Callback (now they are DbCallback and MapsCallback), which can be implemented by different presenters.